moved Begin, Separate and End from page-number.ts to field.ts
